Update http_server.py
Great work! One lesson here is: whatever the type of the file, it can be processed the same way: read as bytes and passed onto the web browser.

Using a "b" in the `open` makes this a little bit easier. This makes Python read the file directly as bytes, so that they don't have to be converted into bytes in a later line.